// graphql/generated/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: { input: any; output: any; }
  /** The `JSONObject` scalar type represents JSON objects as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSONObject: { input: any; output: any; }
  /** The `BigInt` scalar type represents non-fractional signed whole numeric values. */
  Long: { input: any; output: any; }
  /** Mongo object id scalar type */
  ObjectId: { input: any; output: any; }
};
// ...
export type CompleteFrom = {
  __typename?: 'CompleteFrom';
  all?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
  current?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
};
// ...
export type DummyFields = {
  __typename?: 'DummyFields';
  field_number?: Maybe<Scalars['String']['output']>;
  short_name?: Maybe<Scalars['String']['output']>;
};
// ...
export type FieldList = {
  __typename?: 'FieldList';
  complete_from?: Maybe<CompleteFrom>;
  default?: Maybe<Scalars['String']['output']>;
  display_type?: Maybe<Scalars['String']['output']>;
  field_number?: Maybe<Scalars['String']['output']>;
  hide_search?: Maybe<Scalars['Boolean']['output']>;
  instant_facet?: Maybe<Scalars['Boolean']['output']>;
  item_fields?: Maybe<ItemFields>;
  long_description?: Maybe<Scalars['String']['output']>;
  long_name?: Maybe<Scalars['String']['output']>;
  mongo_type?: Maybe<Scalars['String']['output']>;
  notes?: Maybe<Scalars['String']['output']>;
  permitted_values?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
  placeholder?: Maybe<Scalars['String']['output']>;
  public?: Maybe<Scalars['Boolean']['output']>;
  required?: Maybe<Scalars['Boolean']['output']>;
  short_description?: Maybe<Scalars['String']['output']>;
  short_name?: Maybe<Scalars['String']['output']>;
  subfields?: Maybe<Array<Maybe<Subfield>>>;
  weight?: Maybe<Scalars['Int']['output']>;
};
// ...
export type ItemFields = {
  __typename?: 'ItemFields';
  complete_from?: Maybe<CompleteFrom>;
  default?: Maybe<Scalars['String']['output']>;
  display_type?: Maybe<Scalars['String']['output']>;
  field_number?: Maybe<Scalars['String']['output']>;
  instant_facet?: Maybe<Scalars['Boolean']['output']>;
  long_description?: Maybe<Scalars['String']['output']>;
  long_name?: Maybe<Scalars['String']['output']>;
  mongo_type?: Maybe<Scalars['String']['output']>;
  permitted_values?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
  placeholder?: Maybe<Scalars['String']['output']>;
  public?: Maybe<Scalars['Boolean']['output']>;
  required?: Maybe<Scalars['Boolean']['output']>;
  short_description?: Maybe<Scalars['String']['output']>;
  short_name?: Maybe<Scalars['String']['output']>;
  weight?: Maybe<Scalars['Int']['output']>;
};
// ...
export type Subfield = {
  __typename?: 'Subfield';
  complete_from?: Maybe<SubfieldCompleteFrom>;
  default?: Maybe<Scalars['String']['output']>;
  display_type?: Maybe<Scalars['String']['output']>;
  field_number?: Maybe<Scalars['String']['output']>;
  hide_search?: Maybe<Scalars['Boolean']['output']>;
  instant_facet?: Maybe<Scalars['Boolean']['output']>;
  long_description?: Maybe<Scalars['String']['output']>;
  long_name?: Maybe<Scalars['String']['output']>;
  mongo_type?: Maybe<Scalars['String']['output']>;
  permitted_values?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
  placeholder?: Maybe<Scalars['String']['output']>;
  public?: Maybe<Scalars['Boolean']['output']>;
  required?: Maybe<Scalars['Boolean']['output']>;
  short_description?: Maybe<Scalars['String']['output']>;
  short_name?: Maybe<Scalars['String']['output']>;
  weight?: Maybe<Scalars['Int']['output']>;
};
// ...
export type Taxa = {
  __typename?: 'Taxa';
  _id?: Maybe<Scalars['ObjectId']['output']>;
  automatedClassifications?: Maybe<Scalars['Boolean']['output']>;
  complete_entities?: Maybe<Scalars['Boolean']['output']>;
  description?: Maybe<Scalars['String']['output']>;
  dummy_fields?: Maybe<Array<Maybe<DummyFields>>>;
  field_list?: Maybe<Array<Maybe<FieldList>>>;
  namespace?: Maybe<Scalars['String']['output']>;
  weight?: Maybe<Scalars['Int']['output']>;
};